creating output checkers for different types of data.
